pyproject.toml
[project]
name = "qfs"
version = "0.1.0"
description = "qfs height RL training project"
authors = [
    { name = "Nilay Patel", email = "nilaypatel2@gmail.com" }
]
dependencies = [
    "torch>=2.4.0",
    "torchrl==0.6.0",
    "tensordict>=0.5.0",
    "tqdm>=4.66.5",
    "juliacall>=0.9.23",
    "juliapkg>=0.1.14",
    "pyzmq>=26.2.0",
    "msgpack>=1.1.0",
    "pandas>=2.2.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipython>=8.27.0",
    "icecream>=2.1.3",
]

[tool.rye.scripts]
juliad = { call = "rlmath_pt.juliaserver:run" }

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/qfs"]
